move scaling functions to own module
